+++ example/ionic-angular-v3/src/app/tab2/tab2.page.ts
@@ -38,4 +38,8 @@ export class Tab2Page {
       `${Date.now()}: Captured message with added context.`,
     );
   }
+
+  public close(): void {
+    Sentry.close();
+  }
 }