+++ example/ionic-angular-v6/capacitor.config.ts
@@ -3,7 +3,6 @@ import { CapacitorConfig } from '@capacitor/cli';
 const config: CapacitorConfig = {
   appId: 'io.ionic.starter',
   appName: 'ionic-angular',
-  bundledWebRuntime: false,
   plugins: {
     SplashScreen: {
       launchShowDuration: 0,