+++ example/ionic-angular-v6/src/app/tab2/tab2.page.ts
@@ -37,4 +37,8 @@ export class Tab2Page {
       `${Date.now()}: Captured message with added context.`,
     );
   }
+
+  public close(): void {
+    Sentry.close();
+  }
 }