+++ test/wrapper.test.ts
@@ -61,6 +61,7 @@ jest.mock('../src/plugin', () => {
       setUser: jest.fn(() => {
         return;
       }),
+      closeNativeSdk: jest.fn(() => Promise.resolve()),
     },
   };
 });
@@ -624,5 +625,23 @@ describe('Tests Native Wrapper', () => {
 
       expect(SentryCapacitor.fetchNativeDeviceContexts).not.toBeCalled();
     });
+  });
+
+  describe('closeNativeSdk', () => {
+    test('closeNativeSdk calls native bridge', async () => {
+      await NATIVE.closeNativeSdk();
+
+      expect(SentryCapacitor.closeNativeSdk).toBeCalled();
+      expect(NATIVE.enableNative).toBe(false);
+    });
+  });
+
+  test('closeNativeSdk called twice does not throw error.', async () => {
+    logger.debug = jest.fn();
+    await NATIVE.closeNativeSdk();
+    expect(logger.debug).not.toBeCalled();
+
+    await NATIVE.closeNativeSdk();
+    expect(logger.debug).toBeCalledWith(NATIVE._DisabledNativeError);
   });
 });